+++ src/index.js
@@ -9,13 +9,13 @@
 //   document.getElementById('root')
 // );
 
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
- <React.StrictMode>
-  <App />
- </React.StrictMode>
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 );
