+++ src/pages/MyNewToDoList.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const MyNewToDoList = (props) => {
+  return (
+    <div>
+      <h1>New Todo List</h1>
+    </div>
+  );
+};
+
+export default MyNewToDoList;
